@@ -4,9 +4,15 @@
   "description": "A GitHub Octicons icon React component",
   "main": "./lib/index.js",
   "scripts": {
-    "build": "babel src --out-dir lib",
-    "build-demo": "webpack --config=demo/webpack.config.js"
+    "build": "nwb build",
+    "lint": "eslint src tests demo/src",
+    "test": "npm run lint && nwb test",
+    "start": "nwb serve"
   },
+  "files": [
+    "css",
+    "lib"
+  ],
   "repository": {
     "type": "git",
     "url": "https://github.com/insin/react-octicon"
@@ -23,23 +29,16 @@
   },
   "homepage": "https://github.com/insin/react-octicon",
   "dependencies": {
-    "octicons": "^2.2.0"
+    "octicons": "^3.1.0"
   },
   "peerDependencies": {
     "react": ">=0.13.0"
   },
   "devDependencies": {
-    "autoprefixer-loader": "^2.0.0",
-    "babel": "^5.6.14",
-    "babel-core": "^5.6.15",
-    "babel-loader": "^5.2.2",
-    "css-loader": "^0.15.1",
-    "extract-text-webpack-plugin": "^0.8.2",
-    "file-loader": "^0.8.4",
-    "node-libs-browser": "^0.5.2",
-    "react": ">=0.13.0",
-    "style-loader": "^0.12.3",
-    "url-loader": "^0.5.6",
-    "webpack": "^1.10.0"
+    "eslint-config-jonnybuchanan": "^2.0.1",
+    "nwb": "~0.2.0",
+    "react": "^0.14.3",
+    "react-addons-test-utils": "^0.14.3",
+    "react-dom": "^0.14.3"
   }
 }
